src/app/gravity-world/gravity-world.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { GravityWorldComponent } from './gravity-world.component';

describe('Component: GravityWorld', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [GravityWorldComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([GravityWorldComponent],
      (component: GravityWorldComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(GravityWorldComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(GravityWorldComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-gravity-world></app-gravity-world>
  `,
  directives: [GravityWorldComponent]
})
class GravityWorldComponentTestController {
}

